Trilha - Inteligencia Artificial/requirements.txt
ipykernel
openai>=1.0.0
python-dotenv
yfinance
pandas
matplotlib
Pillow
pyaudio
SpeechRecognition
sounddevice
pygame
path
streamlit